[AN] chore: setting xml 삭제
