--- android/app/src/main/java/com/onair/hearit/presentation/main/DrawerClickListener.kt
@@ -1,5 +1,0 @@
-package com.onair.hearit.presentation.main
-
-interface DrawerClickListener {
-    fun openDrawer()
-}
