--- android/app/src/main/java/com/onair/hearit/presentation/setting/SettingComposeFragment.kt
@@ -1,50 +1,0 @@
-package com.onair.hearit.presentation.setting
-
-import android.content.Intent
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import androidx.fragment.app.viewModels
-import androidx.navigation.compose.rememberNavController
-import com.onair.hearit.presentation.login.LoginActivity
-import com.onair.hearit.presentation.main.MainViewModel
-
-class SettingComposeFragment : Fragment() {
-    private val viewModel: SettingViewModel by viewModels { SettingViewModelFactory() }
-
-    private val mainViewModel: MainViewModel by activityViewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View =
-        ComposeView(requireContext()).apply {
-            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
-            setContent {
-                val navController = rememberNavController()
-                SettingNavHost(
-                    navController,
-                    viewModel,
-                    onExitSetting = {
-                        parentFragmentManager.popBackStack()
-                    },
-                    onLogin = {
-                        startActivity(
-                            Intent(
-                                requireContext(),
-                                LoginActivity::class.java,
-                            ),
-                        )
-                    },
-                    onLogout = { mainViewModel.logout() },
-                    onWithdraw = { mainViewModel.withdraw() },
-                )
-            }
-        }
-}
