+++ android/app/src/main/java/com/onair/hearit/presentation/setting/SettingFragment.kt
@@ -1,92 +1,50 @@
 package com.onair.hearit.presentation.setting
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import com.google.firebase.analytics.FirebaseAnalytics
-import com.onair.hearit.analytics.AnalyticsParamKeys.SCREEN_NAME_SETTING
-import com.onair.hearit.databinding.FragmentSettingBinding
-import com.onair.hearit.di.AnalyticsProvider
-import com.onair.hearit.presentation.showToast
-import kotlinx.coroutines.launch
+import androidx.navigation.compose.rememberNavController
+import com.onair.hearit.presentation.login.LoginActivity
+import com.onair.hearit.presentation.main.MainViewModel
 
 class SettingFragment : Fragment() {
-    @Suppress("ktlint:standard:backing-property-naming")
-    private var _binding: FragmentSettingBinding? = null
-    private val binding get() = _binding!!
     private val viewModel: SettingViewModel by viewModels { SettingViewModelFactory() }
+
+    private val mainViewModel: MainViewModel by activityViewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
-    ): View {
-        _binding = FragmentSettingBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(
-        view: View,
-        savedInstanceState: Bundle?,
-    ) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.viewModel = viewModel
-        binding.lifecycleOwner = viewLifecycleOwner
-        setupWindowInsets()
-        setupListener()
-        observeViewModel()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        AnalyticsProvider.get().logEvent(
-            FirebaseAnalytics.Event.SCREEN_VIEW,
-            mapOf(
-                FirebaseAnalytics.Param.SCREEN_NAME to SCREEN_NAME_SETTING,
-                FirebaseAnalytics.Param.SCREEN_CLASS to this::class.simpleName.orEmpty(),
-            ),
-        )
-    }
-
-    private fun setupWindowInsets() {
-        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(0, systemBars.top, 0, 0)
-            insets
-        }
-    }
-
-    private fun setupListener() {
-        binding.ibBack.setOnClickListener {
-            parentFragmentManager.popBackStack()
-        }
-    }
-
-    private fun observeViewModel() {
-        viewLifecycleOwner.lifecycleScope.launch {
-            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                viewModel.userInfo.collect { userInfo ->
-                    binding.userInfo = userInfo
-                }
+    ): View =
+        ComposeView(requireContext()).apply {
+            setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
+            setContent {
+                val navController = rememberNavController()
+                SettingNavHost(
+                    navController,
+                    viewModel,
+                    onExitSetting = {
+                        parentFragmentManager.popBackStack()
+                    },
+                    onLogin = {
+                        startActivity(
+                            Intent(
+                                requireContext(),
+                                LoginActivity::class.java,
+                            ),
+                        )
+                    },
+                    onLogout = { mainViewModel.logout() },
+                    onWithdraw = { mainViewModel.withdraw() },
+                )
             }
         }
-
-        viewModel.toastMessage.observe(viewLifecycleOwner) { resId ->
-            showToast(resId)
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
