+++ android/app/src/main/res/drawable/ic_setting.xml
@@ -5,5 +5,5 @@
     android:viewportHeight="28">
   <path
       android:pathData="M12.628,25.666C12.103,25.666 11.651,25.491 11.273,25.141C10.894,24.791 10.665,24.364 10.587,23.858L10.324,21.933C10.071,21.836 9.833,21.719 9.61,21.583C9.387,21.447 9.168,21.301 8.953,21.146L7.145,21.904C6.659,22.118 6.173,22.137 5.687,21.962C5.201,21.787 4.821,21.476 4.549,21.029L3.178,18.637C2.906,18.19 2.828,17.714 2.945,17.208C3.062,16.702 3.324,16.284 3.733,15.954L5.278,14.787C5.259,14.651 5.249,14.52 5.249,14.393V13.605C5.249,13.479 5.259,13.348 5.278,13.212L3.733,12.045C3.324,11.715 3.062,11.297 2.945,10.791C2.828,10.286 2.906,9.809 3.178,9.362L4.549,6.971C4.821,6.523 5.201,6.212 5.687,6.037C6.173,5.862 6.659,5.882 7.145,6.096L8.953,6.854C9.167,6.698 9.391,6.552 9.624,6.416C9.858,6.28 10.091,6.164 10.324,6.066L10.587,4.141C10.665,3.636 10.893,3.208 11.273,2.858C11.652,2.508 12.104,2.333 12.628,2.333H15.37C15.895,2.333 16.347,2.508 16.727,2.858C17.107,3.208 17.335,3.636 17.412,4.141L17.674,6.066C17.927,6.164 18.165,6.28 18.389,6.416C18.613,6.552 18.832,6.698 19.045,6.854L20.853,6.096C21.34,5.882 21.826,5.862 22.312,6.037C22.798,6.212 23.177,6.523 23.449,6.971L24.82,9.362C25.092,9.809 25.17,10.286 25.053,10.791C24.937,11.297 24.674,11.715 24.266,12.045L22.72,13.212C22.74,13.348 22.749,13.48 22.749,13.606V14.393C22.749,14.52 22.73,14.651 22.691,14.787L24.237,15.954C24.645,16.284 24.908,16.702 25.024,17.208C25.141,17.714 25.063,18.19 24.791,18.637L23.391,21.029C23.119,21.476 22.74,21.787 22.253,21.962C21.767,22.137 21.281,22.118 20.795,21.904L19.045,21.146C18.831,21.301 18.608,21.447 18.374,21.583C18.141,21.719 17.908,21.836 17.674,21.933L17.412,23.858C17.334,24.364 17.106,24.791 16.727,25.141C16.348,25.491 15.896,25.666 15.37,25.666H12.628ZM14.058,18.083C15.185,18.083 16.148,17.684 16.945,16.887C17.742,16.09 18.141,15.127 18.141,14C18.141,12.872 17.742,11.909 16.945,11.112C16.148,10.315 15.185,9.916 14.058,9.916C12.91,9.916 11.943,10.315 11.155,11.112C10.367,11.909 9.973,12.872 9.974,14C9.975,15.127 10.369,16.09 11.156,16.887C11.943,17.684 12.91,18.083 14.058,18.083Z"
-      android:fillColor="#EFF1F2"/>
+      android:fillColor="@color/hearit_gray4"/>
 </vector>
