--- android/app/src/main/res/layout/fragment_setting.xml
@@ -1,143 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="userInfo"
-            type="com.onair.hearit.domain.model.UserInfo" />
-
-        <variable
-            name="viewModel"
-            type="com.onair.hearit.presentation.setting.SettingViewModel" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/hearit_black1">
-
-        <ImageButton
-            android:id="@+id/ib_back"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:background="@color/hearit_gray1"
-            android:contentDescription="@null"
-            android:elevation="1dp"
-            android:src="@drawable/ic_back"
-            app:layout_constraintBottom_toBottomOf="@id/tv_setting_title"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="@id/tv_setting_title" />
-
-        <TextView
-            android:id="@+id/tv_setting_title"
-            style="@style/pretendard_main_title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@color/hearit_gray1"
-            android:gravity="center"
-            android:paddingVertical="16dp"
-            android:text="내정보"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <ImageView
-            android:id="@+id/iv_setting_profile"
-            android:layout_width="120dp"
-            android:layout_height="120dp"
-            android:layout_marginTop="40dp"
-            android:contentDescription="@null"
-            android:scaleType="centerCrop"
-            app:imageUrl="@{userInfo.profileImage}"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_setting_title"
-            tools:src="@mipmap/ic_launcher" />
-
-        <TextView
-            android:id="@+id/tv_setting_nickname"
-            style="@style/pretendard_main_title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="32dp"
-            android:background="@drawable/bg_purple1_radius_24dp"
-            android:gravity="center"
-            android:paddingHorizontal="60dp"
-            android:paddingVertical="12dp"
-            android:text="@{userInfo.nickname}"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/iv_setting_profile"
-            tools:text="닉네임은최대9글자" />
-
-        <com.google.android.material.divider.MaterialDivider
-            android:id="@+id/divider1"
-            android:layout_width="0dp"
-            android:layout_height="1dp"
-            android:layout_marginHorizontal="24dp"
-            android:layout_marginTop="32dp"
-            app:dividerColor="@color/hearit_gray2"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_setting_nickname" />
-
-        <TextView
-            style="@style/pretendard_main_title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="28dp"
-            android:text="@string/setting_notification"
-            android:visibility="gone"
-            app:layout_constraintStart_toStartOf="@id/divider1"
-            app:layout_constraintTop_toBottomOf="@id/divider1" />
-
-        <TextView
-            android:id="@+id/tv_setting_notification_text"
-            style="@style/pretendard_setting_body"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="72dp"
-            android:text="@string/setting_recommend_notification"
-            android:visibility="gone"
-            app:layout_constraintStart_toStartOf="@id/divider1"
-            app:layout_constraintTop_toBottomOf="@id/divider1" />
-
-        <androidx.appcompat.widget.SwitchCompat
-            android:id="@+id/switch_notification"
-            android:layout_width="48dp"
-            android:layout_height="48dp"
-            android:thumb="@drawable/check_switch_thumb"
-            android:visibility="gone"
-            app:layout_constraintBottom_toBottomOf="@id/tv_setting_notification_text"
-            app:layout_constraintEnd_toEndOf="@id/divider1"
-            app:layout_constraintTop_toTopOf="@id/tv_setting_notification_text"
-            app:track="@drawable/check_switch_track" />
-
-        <TextView
-            android:id="@+id/tv_setting_app_version"
-            style="@style/pretendard_body"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="12dp"
-            android:text="@{String.format(@string/setting_app_version, viewModel.appVersion)}"
-            android:textColor="@color/hearit_gray4"
-            app:layout_constraintBottom_toTopOf="@id/tv_setting_app_email"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            tools:text="앱 버전 v1.1.1-DEBUG" />
-
-        <TextView
-            android:id="@+id/tv_setting_app_email"
-            style="@style/pretendard_body"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="80dp"
-            android:text="@string/all_hearit_email"
-            android:textColor="@color/hearit_gray4"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
