+++ android/app/src/main/res/values/font_style.xml
@@ -14,14 +14,7 @@
         <item name="android:textSize">60sp</item>
         <item name="android:textColor">@color/white</item>
     </style>
-
-    <style name="pretendard_drawer">
-        <item name="fontFamily">@font/pretendardmedium</item>
-        <item name="android:includeFontPadding">false</item>
-        <item name="android:textSize">16sp</item>
-        <item name="android:textColor">@color/hearit_gray4</item>
-    </style>
-
+    
     <style name="pretendard_nickname">
         <item name="fontFamily">@font/pretendardbold</item>
         <item name="android:includeFontPadding">false</item>
